Add Escapist Magazine
This one sucked.
If you didn't send an Accept Header (I found), then they returned 500
errors.
But not only that, after the first failed request you went on some sort
of blacklist, and every request, even ones that worked moments ago,
would all return 500s for a while.
Somewhere between hours and days.

There was also a weird list I was on for a while where the requests
worked but every video I requested got me a trailer for the site.
That was weird.