Make Generic More Recursive
So, when it finds an iframe it now goes through the whole plugin process
again.
That means that if there's an iframe to youtube, it will find the
youtube plugin and use it on that iframe.

This means YouTube embeds should get downloaded correctly, but could
also be bad.
It seems like kind of a hacky way to acheive this, but it seems to
work...