New utility: ff-get-flash-video
This utility tries to support the htpc usecase: the video is already
running in a browser window, and user wants to download it instead
for various reasons.

The interface here is typically a remote, not a keyboard. User
basically just pushes the 'Record' button while watching a stream
in browser. There is no place to define file names, so script
stores downloaded files with fixed names bas on source e. g.,
youtube.flv.

To my understanding this cannot really be do browser-independent. So
this is for Firefox (ff).

The scripts basically communicates with firefox in order to get the
url currently played, and then invokes get-flash-video with arguments
describing where to store things.