Use YouTube's rel="canonical"
There are a few different kinds of URLs that resolve to the same page.
In this case, "http://www.youtube.com/embed/CODE" is the same as
"http://www.youtube.com/watch?v=CODE"

So, rather than add a special case for it, I just used the tag in
YouTube's markup that says "I'm the same as this url"

That way the logic is consolidated across the different formats.

I'm almost leaning towards just having the upstream always perform
this... but that might involve a lot more checking than I have time for
right now.