Bad RegEx Looking for Video ID
There was a "video_id=" in http://www.youtube.com/watch?v=9bZkp7q19f0,
but it didn't have an '&' after it as the regex was expecting, so it
overran a and consumed a substantial part of the page.

So, I've changed it.
I don't know if it's not matching now, so it's going to the more
effective fall-through, or it's more correct now, but either way it's
working.