Update Wat.pm
I have been analysing the wat.tv flashplayer and found out, that it takes the time from the http://www.wat.tv/servertime address. The first value is the time in unixtime. This one just needs to be converted to hex and that's it. My greasemonkey script for direct playing the video with a video plugin is working with that.